refactor: split leaf controls from handler containers
