+++ src/pages/json-explorer/features/JsonExplorer/ui/JsonExplorerDataSourceSection.tsx
@@ -1,51 +1,12 @@
-import { ChevronDown, ChevronRight, Database, FileJson, type LucideIcon, Pencil } from 'lucide-react';
+import { ChevronDown, ChevronRight, Database, FileJson, Pencil } from 'lucide-react';
 import { useState } from 'react';
 import type { DataSource } from '../lib/loadTestData';
+import { JsonExplorerDataSourceOption } from './JsonExplorerDataSourceOption';
 
 interface JsonExplorerDataSourceSectionProps {
   dataSource: DataSource;
   changeDataSource: (source: DataSource) => void;
   openCustomJson: () => void;
-}
-
-function DataSourceOption({
-  active,
-  icon: Icon,
-  label,
-  mono = false,
-  source,
-  changeDataSource,
-  openCustomJson,
-}: {
-  active: boolean;
-  icon: LucideIcon;
-  label: string;
-  mono?: boolean;
-  source?: DataSource;
-  changeDataSource: (source: DataSource) => void;
-  openCustomJson: () => void;
-}) {
-  function handleClick() {
-    if (source) {
-      changeDataSource(source);
-      return;
-    }
-
-    openCustomJson();
-  }
-
-  return (
-    <div
-      className={`flex items-center gap-2 px-3 py-1.5 text-2xs cursor-pointer transition-colors ${
-        active ? 'bg-warm-500/20 text-text-primary' : 'text-text-secondary hover:bg-warm-500/10'
-      }`}
-      onClick={handleClick}
-    >
-      <div className={`w-1 h-1 rounded-full shrink-0 ${active ? 'bg-warm-400' : 'bg-text-tertiary'}`} />
-      <Icon size={12} className="shrink-0" />
-      <span className={`${mono ? 'font-mono ' : ''}truncate`}>{label}</span>
-    </div>
-  );
 }
 
 export function JsonExplorerDataSourceSection({
@@ -72,7 +33,7 @@ export function JsonExplorerDataSourceSection({
 
       {isExpanded && (
         <div className="py-1">
-          <DataSourceOption
+          <JsonExplorerDataSourceOption
             active={dataSource === 'test.json'}
             icon={FileJson}
             label="test.json"
@@ -81,7 +42,7 @@ export function JsonExplorerDataSourceSection({
             changeDataSource={changeDataSource}
             openCustomJson={openCustomJson}
           />
-          <DataSourceOption
+          <JsonExplorerDataSourceOption
             active={dataSource === 'test2.json'}
             icon={FileJson}
             label="test2.json"
@@ -90,7 +51,7 @@ export function JsonExplorerDataSourceSection({
             changeDataSource={changeDataSource}
             openCustomJson={openCustomJson}
           />
-          <DataSourceOption
+          <JsonExplorerDataSourceOption
             active={dataSource === 'custom'}
             icon={Pencil}
             label="Custom JSON..."
